Fixed warnings about sign conversions.

+++ MirrorServer/Server/mainwindow.cpp
@@ -84,7 +84,7 @@ void MainWindow::handleFrame(const cv::Mat &matrix) {
     try {
         cv::Mat copy;
         cv::cvtColor(matrix, copy, CV_BGR2RGB);
-        QImage image(copy.data, copy.cols, copy.rows, copy.step, QImage::Format_RGB888);
+        QImage image(copy.data, copy.cols, copy.rows, static_cast<int>(copy.step), QImage::Format_RGB888);
         // Make sure the image fits on screen by scaling it to 500 px high.
         QPixmap pixmap = QPixmap::fromImage(image, Qt::ColorOnly).scaledToHeight(500);
